requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
boto==2.38.0
celery==3.1.17
cloudinary==1.0.21
Django==1.7.4
django-admin-bootstrapped==2.3.2
django-celery==3.1.16
django-filter==0.9.2
django-filters==0.1.0
django-guardian==1.2.5
django-jfu==2.0.9
django-user-roles==0.1.0
djangorestframework==3.0.4
docutils==0.12
elasticsearch==1.4.0
fake-factory==0.4.2
kombu==3.0.26
lxml==3.4.4
milkman==0.4.5
mixer==5.0.7
MySQL-python==1.2.5
netaddr==0.7.14
oauthlib==0.7.2
poster==0.8.1
pycrypto==2.6.1
PyJWT==0.4.1
pylibmc==1.4.1
python-dateutil==2.4.1
python-memcached==1.54
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2014.10
redis==2.10.3
requests==2.5.1
requests-oauthlib==0.4.2
setuptools-git==1.1
simplejson==3.6.5
six==1.9.0
Unirest==1.1.7
urllib3==1.10.2
